Move content of executor to embedded_runtime_version
